requirements.txt
django==3.2.7
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.21.3
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.5